Refactor tests.
- Move test-cases to better places.
